Add Connection class
Now the BasePeer must establish a connection with a PluggableObj
for communicating both messages.

If you want to add another PluggableObj you simply have to implements
the IPluggable interface.

If you want to add another Peer you simply have to extends the
BasePeer class.

To establish a connection, you simply have to create a Connection and
pass to the BasePeer (or its subclass).

The Connection represents the interface for all the communications
between the BasePeer and a PluggableObj.