Update the Logger class, removed the Loggable abstract class.
The Loggable abstract class was removed because the strategy
in which it was used, imposed that all the RunnableInstances
are Loggable and that's not good.

We have only the Logger class that provides the log() method
for logging events.

The Logger has a Formatter scheme (or the default if not provided)
and a Recipient to log events (or the default if not provided).

A class that wants to use a Logger for logging events should simply
call one of the three Logger constructor, get the Logger instace
and log events by calling log() method.