Add Logger functionality to Demo.
Now the messages are logged using the Logger.

There is a problem with the Dashboard's Logger
because the Dashboard is one and the Peer's are many.

The Logger by default expect to respond to a single Peer and
not many Peers.

For this reason the next thing to do is fix this problem to
obtain a smoother program.