Add the queue package
Now the Logger system has been reshaped:

If a class wants to use a logger, that class must implement
the QueueListener interface.

You have to create the specific Queue (implementation of IQueue)
and you have to register a new instance of Logger to the queue
and the BasePeer subclass instance to the same Queue.

In this way when the BasePeer subclass instance puts a message to be
logged, the Logger that is waiting for a message to be logged can log
it.

This is a more efficient approach since when the Logger has no work to do
it blocks without consuming CPU cycles.